Isolated channel types + creation of contact job is part of same transaction
