+++ hrm-service/src/contact-job/contact-job-data-access.ts
@@ -1,3 +1,4 @@
+import { ITask } from 'pg-promise';
 import { db, pgp } from '../connection-pool';
 import { Contact } from '../contact/contact-data-access';
 import { COMPLETE_JOB_SQL, PULL_DUE_JOBS_SQL } from './sql/contact-job-sql';
@@ -66,22 +67,23 @@ export const completeContactJob = async (
 
 /**
  * Add a new job to be completed to the ContactJobs queue
- * @param job
+ * Requires tx: ITask to make the creation of the job part of the same transaction
  */
-export const createContactJob = async (job: Omit<ContactJob, 'id'>): Promise<void> => {
+export const createContactJob = (tx: ITask<{}>) => async (
+  job: Omit<ContactJob, 'id'>,
+): Promise<void> => {
   const contact = job.resource;
-  await db.task(tx => {
-    const insertSql = pgp.helpers.insert(
-      {
-        requested: new Date().toISOString(),
-        jobType: job.jobType,
-        contactId: contact.id,
-        accountSid: contact.accountSid,
-        additionalPayload: job.additionalPayload,
-      },
-      null,
-      'ContactJobs',
-    );
-    return tx.none(insertSql);
-  });
+  const insertSql = pgp.helpers.insert(
+    {
+      requested: new Date().toISOString(),
+      jobType: job.jobType,
+      contactId: contact.id,
+      accountSid: contact.accountSid,
+      additionalPayload: job.additionalPayload,
+    },
+    null,
+    'ContactJobs',
+  );
+
+  return tx.none(insertSql);
 };
